[WOR-1560] Add StairwayLoggingHook for reuse
This StairwayHook is similar to existing MonitoringHook (not a Spring component, clients can construct it directly when initializing their Stairways).
This minimizes code duplication across services and removes a potential for autowiring conflicts between services that both define their own StairwayLoggingHooks as Spring components (ex. WSM and LZS, the former pulls in the latter).