[PF-2953] Use Spring dependency manager
Previously, this was disabled because it generated invalid POM and broke publishing to mavenLocal, which is useful for testing.

This is no longer the case: I verified that when bringing Spring dependency management back in at the latest release I could still publish to mavenLocal and pull in the resulting artifact into another service.

As a result, many places where we previously versioned dependencies can now derive their versions from the dependency manager.

In many places, we were defining dependencies redundantly: we're pulling in a number of Spring starters which expose more transitive dependencies for our usage than we realized.

I did not upgrade any versions to keep the scope narrow: I'll leave that to Dependabot.